Add vehicle_signal_specification submodule
VSS itself was not added until now but it is of course needed. Since
some required files (*.cnative) are built from
vehicle_signal_specification, and some of vss-tools are required to
prepare those file, therefore of course vehicle_signal_specification
itself needs to be part of this "combined" demonstration repository, and
its version should be controlled so that the combined build is
fully repeatable.