docker/ : add statestorage container
Statestorage is a run-once executable to prepare a database.  In other
words it is not a daemon and therefore not like the other containers.

However, it processes files in the shared /data directory and needs
to be run before others, so it seemed very easy to just set it
up in the exact same structure as the others more long-lived processes.