data/ovds_simulator_input.db committed to git
The simulation data is committed to git to make a livesim component demo
also fully repeatable

This real-world data was collected with user consent, but in this copy
of the db, the VIN was changed to ANON_VIN.  It is arbitrary anyways,
for demo purposes.  In other words, make sure to use ANON_VIN when
running the livesim component with this database file.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>