sqlitebrowser container
Usually it's more convenient to just install and run sqlitebrowser
locally, but it was exceptionally easy to add this to the structure we
already have, so why not.  The container, which is identical to
linuxserver/sqlitebrowser on the Docker hub, exposes the GUI of the
sqlite browser to a web interface on localhost port 3001 here (3000
originally), via guacamole (RDP-to-Web bridge).

http://localhost:3001/